+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.6+2
+
+- Screenshot uploader improvements and fixes
+
+## 0.0.6+1
+
+- Screenshot uploader improvements and fixes
+
 ## 0.0.5
 
 - Screenshot uploader improvements and fixes